Ignore comments in lines connected by a continuation character (#828)
Currently, in the configuration file, if a comment follows a line continuation character, the comment and all the subsequent lines are ignored. With this change the parser just ignores the comments, but not the lines which follow after the comment. So the continuation only applies to the first line after the continuation character which is not a comment.